fixed compiler warning
https://gitlab.inria.fr/core-math/core-math/-/jobs/6792821

+++ src/binary64/lgamma/check_special.c
@@ -341,7 +341,8 @@ static void scan_consecutive(int64_t n, double x){
     n -= jmax;
     x += jmax * ldexp (1.0, e - 53);
   }
-  printf ("checked %llu values, expensive checks %lu\n", n0, tested);
+  printf ("checked %lu values, expensive checks %lu\n",
+          (unsigned long) n0, tested);
 }
 
 int
